chore:move the ui helpers to ui/helpers
