+++ app/src/main/java/be/scri/helpers/CommonsConstants.kt
@@ -14,6 +14,4 @@ const val KEY_COLOR = "key_color"
 const val BACKGROUND_COLOR = "background_color"
 const val PRIMARY_COLOR = "primary_color_2"
 const val ACCENT_COLOR = "accent_color"
-const val USE_ENGLISH = "use_english"
-const val WAS_USE_ENGLISH_TOGGLED = "was_use_english_toggled"
 const val IS_USING_SYSTEM_THEME = "is_using_system_theme"
