+++ app/src/main/java/be/scri/helpers/Constants.kt
@@ -15,7 +15,6 @@ const val MAX_KEYS_PER_MINI_ROW = 9
 // Shared preferences.
 const val VIBRATE_ON_KEYPRESS = "vibrate_on_keypress"
 const val SHOW_POPUP_ON_KEYPRESS = "show_popup_on_keypress"
-const val DARK_THEME = "dark_theme"
 const val PERIOD_ON_DOUBLE_TAP = "period_on_double_tap"
 const val PERIOD_AND_COMMA = "period_and_comma"
 const val DISABLE_ACCENT_CHARACTER = "disable_accent_character"
