--- app/src/main/java/be/scri/views/CustomDividerItemDecoration.kt
@@ -1,46 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-import android.graphics.Canvas
-import android.graphics.Rect
-import android.graphics.drawable.Drawable
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-
-/**
- * A custom divider for use in recycle views.
- */
-class CustomDividerItemDecoration(
-    private val drawable: Drawable,
-    private val width: Int,
-    private val marginLeft: Int,
-    private val marginRight: Int,
-) : RecyclerView.ItemDecoration() {
-    override fun onDraw(
-        canvas: Canvas,
-        parent: RecyclerView,
-        state: RecyclerView.State,
-    ) {
-        val left = parent.paddingLeft + marginLeft
-        val right = parent.width - parent.paddingRight - marginRight
-
-        val childCount = parent.childCount
-        for (i in 0 until childCount - 1) {
-            val child = parent.getChildAt(i)
-            val params = child.layoutParams as RecyclerView.LayoutParams
-            val top = child.bottom + params.bottomMargin
-            val bottom = top + width
-
-            drawable.setBounds(left, top, right, bottom)
-            drawable.draw(canvas)
-        }
-    }
-
-    override fun getItemOffsets(
-        outRect: Rect,
-        view: View,
-        parent: RecyclerView,
-        state: RecyclerView.State,
-    ) {
-        outRect.set(0, 0, 0, width)
-    }
-}
