--- app/src/main/res/layout/activity_main.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:theme="@style/Theme.AppCompat.Light.NoActionBar"
-    android:layout_height="match_parent">
-    <!-- Hint Container -->
-    <include
-        android:id="@+id/hint_layout"
-        layout="@layout/hint_layout"
-        android:layout_marginStart="16dp"
-        android:layout_marginEnd="16dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:visibility="gone" />
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/view_pager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-    <FrameLayout
-        android:id="@+id/fragment_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_above="@id/bottom_navigation"
-        app:layout_anchor="@+id/view_pager"
-        app:layout_anchorGravity="center" />
-    <androidx.compose.ui.platform.ComposeView
-        android:id="@+id/compose_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_above="@id/bottom_navigation"
-        app:layout_anchor="@+id/view_pager"
-        app:layout_anchorGravity="center" />
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/bottom_navigation"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_gravity="bottom"
-        app:itemBackground="@color/nav_bar_color"
-        app:itemIconTint="@drawable/nav_bar_selector"
-        app:itemTextColor="@drawable/nav_bar_selector"
-        app:menu="@menu/menu_navigation_bottom" />
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
