--- app/src/main/res/layout/card_view_with_image.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-    <LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:padding="16dp">
-
-        <ImageView
-            android:id="@+id/imgView1"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:layout_marginEnd="12dp"
-            tools:srcCompat="@tools:sample/avatars" />
-
-        <TextView
-            android:id="@+id/tvText"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="Get the code on GitHub"
-            android:autoSizeTextType="uniform"
-            android:autoSizeMinTextSize="15sp"
-            android:autoSizeMaxTextSize="24sp"
-            android:autoSizeStepGranularity="2sp"
-            android:textStyle="normal" />
-
-        <ImageView
-            android:id="@+id/imgView2"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:layout_marginEnd="8dp"
-            tools:srcCompat="@tools:sample/avatars" />
-
-    </LinearLayout>
