--- app/src/main/res/layout/fragment_settings.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/settings_constraint_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/you_background_color">
-
-
-        <LinearLayout
-            android:id="@+id/settings_holder"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:padding="10dp"
-            android:orientation="vertical">
-
-            <TextView
-                android:layout_marginTop="16dp"
-                android:id="@+id/languageTitle"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingTop="10dp"
-                android:paddingStart="10dp"
-                android:paddingEnd="10dp"
-                android:paddingBottom="10dp"
-                android:text="@string/app.settings.menu.title"
-                android:textColor="@color/app_text_color"
-                android:textSize="20sp"
-                android:textStyle="bold" />
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/recyclerViewSettings"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="8dp"
-                android:layout_marginEnd="8dp"
-                android:background="@drawable/rounded_all_corners"
-               />
-
-            <TextView
-                android:id="@+id/selectLanguage"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingTop="10dp"
-                android:paddingEnd="12dp"
-                android:paddingStart="10dp"
-                android:paddingBottom="10dp"
-                android:text="@string/app.settings.keyboard.title"
-                android:textColor="@color/app_text_color"
-                android:textSize="20sp"
-                android:textStyle="bold"
-                android:visibility="invisible" />
-
-            <RelativeLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-
-                <Button
-                    android:layout_marginEnd="16dp"
-                    android:layout_marginBottom="16dp"
-                    android:id="@+id/btnInstall"
-                    android:layout_width= "600dp"
-                    android:layout_height="80dp"
-                    android:strokeColor="@color/button_outline_color"
-                    android:layout_marginStart="13dp"
-                    android:textStyle="bold"
-                    android:textSize="20sp"
-                    android:strokeWidth="@dimen/tiny_margin"
-                    android:textColor="@color/button_text_color"
-                    android:background="@drawable/rounded_button"
-                    android:padding="10dp"
-                    android:textAllCaps="false"
-                    android:text="@string/app.settings.button_install_keyboards" />
-
-
-                <androidx.recyclerview.widget.RecyclerView
-                    android:id="@+id/recyclerView2"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="8dp"
-                    android:layout_marginEnd="8dp"
-                    android:layout_marginBottom="50dp"
-                    android:background="@drawable/rounded_all_corners"
-
-                    />
-            </RelativeLayout>
-
-        </LinearLayout>
-    </ScrollView>
