--- app/src/main/res/layout/hint_layout.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    app:cardCornerRadius="8dp"
-    app:cardElevation="4dp"
-    android:background="@color/card_view_color"
-    app:cardBackgroundColor="@color/card_view_color">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:padding="10dp">
-
-        <ImageView
-            android:contentDescription="Light bulb icon"
-            android:id="@+id/lightbulb_icon"
-            android:layout_width="38dp"
-            android:layout_height="38dp"
-            android:layout_gravity="center_vertical"
-            android:layout_marginEnd="16dp"
-            android:src="@drawable/light_bulb_icon"
-            app:tint="@color/hint_text_color"
-            />
-
-        <TextView
-            android:id="@+id/hint_text"
-            android:layout_width="0dp"
-            android:textColor="@color/app_text_color"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_vertical"
-            android:layout_weight="1" />
-
-        <Button
-            android:id="@+id/hint_ok_button"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            android:layout_gravity="center_vertical"
-            android:text="@string/hint_ok_button_text"
-            android:textColor="@android:color/white"
-            android:background="@drawable/button_background_rounded"
-            android:backgroundTint="@color/nav_bar_selected_color"
-            />
-
-    </LinearLayout>
-
-</androidx.cardview.widget.CardView>
