--- app/src/main/res/menu/menu_navigation_bottom.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/installation"
-        android:title="@string/app.installation.title"
-        android:icon="@drawable/material_keyboard"/>
-    <item
-        android:id="@+id/settings"
-        android:title="@string/app.settings.title"
-        android:icon="@drawable/material_settings"/>
-
-    <item
-        android:id="@+id/info"
-        android:title="@string/app.about.title"
-        android:icon="@drawable/material_info"/>
-
-</menu>
