--- app/src/main/res/values/ids.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <!-- Commons ids -->
-    <item name="properties_size" type="id" />
-    <item name="properties_direct_children_count" type="id" />
-    <item name="properties_file_count" type="id" />
-    <item name="properties_last_modified" type="id" />
-    <item name="properties_md5" type="id" />
-
-</resources>
